requirements.txt
# Core
pandas
numpy
joblib
scikit-learn

# Transcription
openai-whisper
ffmpeg-python

# API requests
requests